requirements.txt
# Core dependencies
pandas==2.1.0
numpy==1.24.3
pyarrow==14.0.1
datasets==2.14.0
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.18.1

# ML/Stats
scikit-learn==1.3.0
scipy==1.11.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
jupyter==1.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2